Updated tests for CertificateDatesService
